New Update
we have fixed the Menu so now you can use the arrows to select what you want in the menu if it dosen't select anything just type e and it will work.

we have also added whole new enemy with 2 different attacks and much more.

we'll be updating it soon again :)