New update
added qurality sittings and grafic sittings to the game.

but we did not really need so i disableed it.

chanced the Leaderboard disgn a little.

added a main menu and made sure that ones you have confirmed your name you cannot confirm it twice as it would take you to the leaderboard after confirming it.

fixed a few other issues.

we are aware about small lag issues but i'll try to fix it next time :)